move isolate addition to otu.isolate
