+++ ref_builder/cli/main.py
@@ -24,13 +24,15 @@ from ref_builder.options import (
 )
 from ref_builder.otu.create import create_otu
 from ref_builder.otu.update import (
-    add_and_name_isolate,
-    add_genbank_isolate,
-    add_unnamed_isolate,
     auto_update_otu,
     exclude_accessions_from_otu,
     promote_otu_accessions,
     set_representative_isolate,
+)
+from ref_builder.otu.isolate import (
+    add_genbank_isolate,
+    add_unnamed_isolate,
+    add_and_name_isolate,
 )
 from ref_builder.otu.modify import add_segments_to_plan, resize_monopartite_plan
 from ref_builder.otu.utils import RefSeqConflictError
